pyproject.toml
[project]
name = "agent-sanbox"
version = "0.1.0"
description = "Reusable workflow and agent sandbox engine for digital twins"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Richard Gyiko" }]
dependencies = [
    "jsonschema>=4.25.1",
    "opentelemetry-api>=1.27.0",
    "pyyaml>=6.0.2",
    "referencing>=0.36.2",
]

[project.scripts]
agent-sandbox = "agent_sandbox.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_sandbox"]

[tool.hatch.build.targets.wheel.force-include]
"src/agent_sandbox/resources" = "agent_sandbox/resources"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "TCH", "PTH", "PL", "RUF"]
ignore = ["PLC0415", "PLR2004", "PLR0911", "PLR0912", "PLR0913", "PLR0915"]

[tool.ruff.lint.isort]
known-first-party = ["agent_sandbox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short", "--strict-markers"]
filterwarnings = ["ignore::DeprecationWarning"]
